requirements.txt
appdirs==1.4.4
asgiref==3.4.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
distlib==0.3.2
Django==3.2.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
filelock==3.0.12
idna==3.2
Pillow==8.2.0
pipdeptree==2.0.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
svgwrite==1.4.1
urllib3==1.26.6
